DewalGautam/requirements.txt
requests
aws-cdk-lib==2.208.0
constructs>=10.0.0,<11.0.0
